Fix problem with error returns in marpa_r_earley_set_value*() methods
Since -2 was a valid "integer value", these methods could return
it on success.  But these methods did not clear the error code to
MARPA_ERR_NONE, so that, on a return value of -2, there was no way to
distinguish success from failure.  These methods are now fixed to clear
the error code to MARPA_ERR_NONE.